Finished dividing code into classes
There is a method last in main.java that returns the IP-address of the phone. This might come in handy maybe and is therefore not deleted.

Other than that the code is mostly finished. It might need change to the structure if there is a easier way of doing the same thing.