Bug ID_MF5 almost solved
Createing new Thread already in MainMenu Activity.

All games will use this Thread and this thread will also listen to if the Serve connection is lost.

The problem that remains is when the User is redirected to MainFrame then the method onResume() is called. The Applications should start fresh and forget all information.

Also this has to be tested if it works when playing the game. It should.